pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "velocitytree"
version = "0.1.0"
description = "A Python tool to streamline developer workflows by managing project structure, context, and integrating AI assistance"
readme = "README.md"
authors = [
    {name = "Guntram Bechtold", email = "your.email@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Version Control",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "gitpython>=3.1.0",
    "jinja2>=3.0.0",
    "pyyaml>=5.4.0",
    "toml>=0.10.0",
    "openai>=0.27.0",
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    "pathspec>=0.9.0",
    "aiofiles>=0.8.0",
    "aiohttp>=3.8.0",
    "pydantic>=1.9.0",
    "prompt-toolkit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/gbechtold/Velocitytree"
Repository = "https://github.com/gbechtold/Velocitytree"
Issues = "https://github.com/gbechtold/Velocitytree/issues"
Documentation = "https://velocitytree.readthedocs.io"

[project.scripts]
velocitytree = "velocitytree.cli:main"
vtree = "velocitytree.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "pytest-asyncio>=0.18.0",
    "pytest-timeout>=2.0.0",
    "pytest-mock>=3.6.0",
    "flake8>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.910",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "pre-commit>=2.15.0",
    "tox>=3.24.0",
    "twine>=3.4.0",
    "bumpversion>=0.6.0",
]

[tool.setuptools_scm]
write_to = "velocitytree/_version.py"

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]

[tool.coverage.run]
branch = true
source = ["velocitytree"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]